refactor: move guardrails extension files
