+++ extensions/guardrails/components/pattern-editor.ts
