+++ extensions/guardrails/hooks/permission-gate/dangerous-commands.test.ts
