+++ extensions/guardrails/hooks/permission-gate/dangerous-commands.ts
