+++ extensions/guardrails/lib/timing.ts
