+++ tsconfig.json
@@ -10,6 +10,6 @@
     "resolveJsonModule": true,
     "noEmit": true
   },
-  "include": ["src/**/*", "tests/**/*", "vitest.config.ts"],
+  "include": ["src/**/*", "extensions/**/*", "tests/**/*", "vitest.config.ts"],
   "exclude": ["node_modules"]
 }
