+++ vitest.config.ts
@@ -14,7 +14,11 @@ export default defineConfig({
   },
   test: {
     environment: "node",
-    include: ["src/**/*.test.ts", "tests/**/*.test.ts"],
+    include: [
+      "src/**/*.test.ts",
+      "extensions/**/*.test.ts",
+      "tests/**/*.test.ts",
+    ],
     setupFiles: ["./tests/vitest.setup.ts"],
     mockReset: true,
   },
